+++ src/main/java/kitchenpos/menu/application/validator/MenuValidator.java
@@ -1,0 +1,8 @@
+package kitchenpos.menu.application.validator;
+
+import kitchenpos.menu.domain.Menu;
+
+public interface MenuValidator {
+
+    void validate(final Menu menu);
+}
