--- src/main/java/kitchenpos/menu/domain/MenuValidator.java
@@ -1,6 +1,0 @@
-package kitchenpos.menu.domain;
-
-public interface MenuValidator {
-
-    void validate(final Menu menu);
-}
