+++ src/main/java/kitchenpos/order/application/validator/OrderValidator.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.application.validator;
+
+import kitchenpos.order.domain.Order;
+
+public interface OrderValidator {
+
+    void validate(final Order order);
+}
