+++ src/main/java/kitchenpos/order/application/validator/TableGroupValidator.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.application.validator;
+
+import kitchenpos.order.domain.TableGroup;
+
+public interface TableGroupValidator {
+
+    void validate(final TableGroup tableGroup);
+}
