+++ src/main/java/kitchenpos/order/application/validator/TableUngroupValidator.java
@@ -1,4 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.application.validator;
+
+import kitchenpos.order.domain.TableGroup;
 
 public interface TableUngroupValidator {
