--- src/main/java/kitchenpos/order/domain/OrderValidator.java
@@ -1,6 +1,0 @@
-package kitchenpos.order.domain;
-
-public interface OrderValidator {
-
-    void validate(final Order order);
-}
