--- src/main/java/kitchenpos/order/domain/TableGroupValidator.java
@@ -1,6 +1,0 @@
-package kitchenpos.order.domain;
-
-public interface TableGroupValidator {
-
-    void validate(final TableGroup tableGroup);
-}
